tests move close to implementation
