+++ src/pages/pokemon-details/pokemon-image.component.tsx
@@ -1,7 +1,7 @@
+import React from 'react';
 import { Box } from '@mui/material';
 import { Paper } from '@mui/material';
 import { styled } from '@mui/material/styles';
-import React from 'react';
 
 interface Props {
   image: string;
